ast: Fix NONE node escaping parse_block_body
If no statement is specified after a control call, ex.

!repeat(4)  <-- NOTE: \n is an implicit command ending in this case.
print hello

OR

!repeat(4); print hello

...the control call's statement parameter is populated by a NONE-type
node, corresponding to a command separator. This is not a valid
statement, and so raises an InternalInterpreterError when interpreted.

This stems from wrongly considering command-separator tokens valid
statements in the parser. So, remove them from ast.parse_statement and
check separately.