interpreter: Add base call context
In order to simplify extension of the execution context, there is now
no difference between being "in" and "out" of a call. By adding a
toplevel call context object (a call with name __main__ and no
arguments), the interpreter is now always "in" a call.

Also, fix an oversight in Interpreter.interpret_ast that caused
updates to Interpreter.init_context to not affect interpretation
of full AST objects.